Move Iceberg MV with PTF tests to base class
Move tests of Iceberg materialized views based on table functions to the
Iceberg materialized views base test class.
